all the code i write.
it spend me about two days to do this with the help of chatgpt. This is a easy try of using what i learned from class into code, just like a homework.
the ,ost difficult part is how to compare different date. It is really not a easy thing to handle time series.